requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-elasticsearch-dsl==7.2.1
djangorestframework==3.12.4
dnspython==2.1.0
elasticsearch==7.15.2
elasticsearch-dsl==7.4.0
eventlet==0.33.0
flower==1.0.0
greenlet==1.1.2
hiredis==2.0.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==4.6.11
msgpack==1.0.3
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
